@@ -1,14 +1,13 @@
 package vega.financial.calculator;
 
 import org.assertj.core.data.Offset;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
 
 import org.assertj.core.api.Assertions;
 import vega.financial.StockOption.OptionType;
 
-@RunWith(SpringRunner.class)
+@SpringBootTest
 public class TestSpotFinder {
 
     @Test
